+++ spec/integration/rest/conversations/v1/user.spec.js
@@ -73,7 +73,10 @@ describe('User', function() {
           'is_notifiable': null,
           'date_created': '2019-12-16T22:18:37Z',
           'date_updated': '2019-12-16T22:18:38Z',
-          'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+          'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'links': {
+              'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+          }
       };
 
       holodeck.mock(new Response(201, body));
@@ -126,7 +129,10 @@ describe('User', function() {
           'is_notifiable': null,
           'date_created': '2019-12-16T22:18:37Z',
           'date_updated': '2019-12-16T22:18:38Z',
-          'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+          'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'links': {
+              'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+          }
       };
 
       holodeck.mock(new Response(200, body));
@@ -214,7 +220,10 @@ describe('User', function() {
           'is_notifiable': null,
           'date_created': '2019-12-16T22:18:37Z',
           'date_updated': '2019-12-16T22:18:38Z',
-          'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+          'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'links': {
+              'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+          }
       };
 
       holodeck.mock(new Response(200, body));
@@ -253,7 +262,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2019-12-16T22:18:37Z',
                   'date_updated': '2019-12-16T22:18:38Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               },
               {
                   'sid': 'USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -267,7 +279,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2020-03-24T20:38:21Z',
                   'date_updated': '2020-03-24T20:38:21Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               }
           ]
       };
@@ -300,7 +315,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2019-12-16T22:18:37Z',
                   'date_updated': '2019-12-16T22:18:38Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               },
               {
                   'sid': 'USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -314,7 +332,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2020-03-24T20:38:21Z',
                   'date_updated': '2020-03-24T20:38:21Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               }
           ]
       };
@@ -352,7 +373,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2019-12-16T22:18:37Z',
                   'date_updated': '2019-12-16T22:18:38Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               },
               {
                   'sid': 'USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -366,7 +390,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2020-03-24T20:38:21Z',
                   'date_updated': '2020-03-24T20:38:21Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               }
           ]
       };
@@ -419,7 +446,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2019-12-16T22:18:37Z',
                   'date_updated': '2019-12-16T22:18:38Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               },
               {
                   'sid': 'USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -433,7 +463,10 @@ describe('User', function() {
                   'is_notifiable': null,
                   'date_created': '2020-03-24T20:38:21Z',
                   'date_updated': '2020-03-24T20:38:21Z',
-                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
+                  'url': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'links': {
+                      'user_conversations': 'https://conversations.twilio.com/v1/Users/USaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conversations'
+                  }
               }
           ]
       };